use int32 instead int.

+++ Sources/EstruturaDeDados/DistributionSort.swift
@@ -1,6 +1,6 @@
 //
 //  DistributionSort.swift
-//  
+//
 //
 //  Created by Emanuel on 03/07/24.
 //
@@ -8,31 +8,31 @@
 import Foundation
 
 
-public func distributionSort(v: inout [Int]) {
+public func distributionSort(v: inout [Int32]) {
     distributionSort(v: &v, n: v.count)
 }
 
-func distributionSort(v: inout [Int], n: Int) {
-    let s = v.min()!
-    let b = v.max()!
+func distributionSort(v: inout [Int32], n: Int) {
+    let s = Int(v.min()!)
+    let b = Int(v.max()!)
     var c: [Int] = []
     var w: [Int] = []
     for _ in 0..<(b - s + 2) {
         c.append(0)
     }
     for i in 0..<n {
-        c[v[i] - s + 1] = c[v[i] - s + 1] + 1
+        c[Int(v[i]) - s + 1] = c[Int(v[i]) - s + 1] + 1
         w.append(0)
     }
     for i in 1..<(b - s + 2) {
         c[i] += c[i - 1]
     }
     for i in 0..<n {
-        let d = v[i] - s + 1
-        w[c[d] - 1] = v[i]
+        let d = Int(v[i]) - s + 1
+        w[c[d] - 1] = Int(v[i])
         c[d] -= 1
     }
     for i in 0..<n {
-        v[i] = w[i]
+        v[i] = Int32(w[i])
     }
 }
